Split violations_util into focused submodules (quality-loop round 2)
