--- crates/keel-enforce/src/violations_util.rs
@@ -1,945 +1,0 @@
-/// Compute the (plain, disambiguated) hash pair for a definition — the two
-/// identities `keel map` may have stored for it (map salts with the file
-/// path on collision). Single source of truth for "does this def match its
-/// stored node", shared by the engine's hash sync and progressive adoption.
-pub fn definition_hashes(
-    def: &keel_parsers::resolver::Definition,
-    file_path: &str,
-) -> (String, String) {
-    // Normalize the body once and reuse it for both identities — routing
-    // through `Definition::{hash, hash_disambiguated}` would normalize twice.
-    let doc = def.docstring.as_deref().unwrap_or("");
-    let body = def.body_for_hash();
-    (
-        keel_core::hash::compute_hash(&def.signature, &body, doc),
-        keel_core::hash::compute_hash_disambiguated(&def.signature, &body, doc, file_path),
-    )
-}
-
-/// Compute one disambiguated hash for a definition under an arbitrary salt.
-///
-/// [`definition_hashes`] covers the two identities `keel map` assigns (plain,
-/// and file-path-salted). A file holding three or more identical same-named
-/// definitions needs more than two distinct identities, so the engine's
-/// re-baseline walks an ordinal (`"<file>#2"`, `"<file>#3"`, …) through this.
-/// Off the hot path — it re-normalizes the body — and only reached once a
-/// collision is already proven.
-pub fn definition_hash_salted(def: &keel_parsers::resolver::Definition, salt: &str) -> String {
-    keel_core::hash::compute_hash_disambiguated(
-        &def.signature,
-        &def.body_for_hash(),
-        def.docstring.as_deref().unwrap_or(""),
-        salt,
-    )
-}
-
-/// Whether a stored node's hash matches the definition under either of the
-/// two identities from [`definition_hashes`].
-pub fn node_hash_matches(
-    node: &keel_core::types::GraphNode,
-    def: &keel_parsers::resolver::Definition,
-    file_path: &str,
-) -> bool {
-    let (hash, hash_d) = definition_hashes(def, file_path);
-    node.hash == hash || node.hash == hash_d
-}
-
-/// Pair a freshly parsed definition with the stored node that IS it, or
-/// `None` when the pairing is genuinely undecidable.
-///
-/// Every check that compares a parse against the graph needs this, and each
-/// one used to inline its own version. The strategy, in order:
-///
-/// 1. **Hash evidence wins.** A same-named node whose hash matches under
-///    either identity from [`definition_hashes`] is this definition, however
-///    many same-named siblings surround it.
-/// 2. **Otherwise the name must be unique on BOTH sides** — exactly one
-///    definition in the parse and exactly one node in the store. A repeated
-///    name with no hash evidence is a coin flip (a free `search_graph` beside
-///    a `search_graph` method), and comparing the wrong pair manufactures
-///    findings out of nothing.
-/// 3. **Otherwise refuse.** Undecidable is a real answer; guessing is not.
-///
-/// A signature change moves the hash, so a *changed* definition never has
-/// evidence under rule 1 — for repeated names that is precisely the case rule
-/// 2 must refuse.
-pub fn bind_to_node<'a>(
-    def: &keel_parsers::resolver::Definition,
-    file: &keel_parsers::resolver::FileIndex,
-    nodes: &'a [keel_core::types::GraphNode],
-) -> Option<&'a keel_core::types::GraphNode> {
-    let candidates: Vec<&keel_core::types::GraphNode> =
-        nodes.iter().filter(|n| n.name == def.name).collect();
-    if let Some(matched) = candidates
-        .iter()
-        .find(|n| node_hash_matches(n, def, &file.file_path))
-    {
-        return Some(matched);
-    }
-    let unique_in_parse = file
-        .definitions
-        .iter()
-        .filter(|d| d.name == def.name)
-        .count()
-        == 1;
-    (candidates.len() == 1 && unique_in_parse).then(|| candidates[0])
-}
-
-/// Check if a file is a benchmark file — harness-invoked, like tests, so
-/// dead-code analysis doesn't apply.
-pub fn is_bench_file(path: &str) -> bool {
-    let normalized = path.replace('\\', "/");
-    normalized.starts_with("benches/") || normalized.contains("/benches/")
-}
-
-/// Check if a file is a declaration/stub file: signature-only by definition,
-/// so docstring and body-level checks don't apply (`.pyi` stubs, `.d.ts`
-/// declarations).
-pub fn is_stub_file(path: &str) -> bool {
-    path.ends_with(".pyi")
-        || path.ends_with(".d.ts")
-        || path.ends_with(".d.mts")
-        || path.ends_with(".d.cts")
-}
-
-/// True when `normalized` is a Cargo target ROOT under `dir/` (at the repo
-/// root or under any crate): directly `dir/<name>.rs`, or the
-/// `dir/<name>/main.rs` form of a multi-file target. Sibling files of a
-/// multi-file target (`dir/<name>/util.rs`) are modules of that target — they
-/// share its namespace and are NOT roots.
-fn target_root_under(normalized: &str, dir: &str) -> bool {
-    let prefix = format!("{dir}/");
-    let rest = if let Some(r) = normalized.strip_prefix(&prefix) {
-        r
-    } else if let Some(pos) = normalized.find(&format!("/{prefix}")) {
-        &normalized[pos + 1 + prefix.len()..]
-    } else {
-        return false;
-    };
-    match rest.split('/').collect::<Vec<_>>().as_slice() {
-        [file] => file.ends_with(".rs"),
-        [_, main] => *main == "main.rs",
-        _ => false,
-    }
-}
-
-/// True for a Cargo-recognized binary/build-script compilation root:
-/// crate-root `build.rs`, `src/main.rs`, a `src/bin` target root, or an
-/// `examples` target root. `.rs` only — the layout conventions are Cargo's,
-/// and a `src/bin/cli.ts` is an ordinary module that can import its siblings.
-fn is_cargo_binary_root(path: &str) -> bool {
-    let normalized = path.replace('\\', "/");
-    if !normalized.ends_with(".rs") {
-        return false;
-    }
-    // `src/build.rs` is a plain module, not a build script: a crate-root
-    // build.rs never sits under src/.
-    let crate_root_build = normalized == "build.rs"
-        || (normalized.ends_with("/build.rs")
-            && !normalized.ends_with("/src/build.rs")
-            && normalized != "src/build.rs");
-    crate_root_build
-        || normalized == "src/main.rs"
-        || normalized.ends_with("/src/main.rs")
-        || target_root_under(&normalized, "src/bin")
-        || target_root_under(&normalized, "examples")
-}
-
-/// True when two files are compiled as SEPARATE Cargo units — a build script,
-/// a `src/main.rs`, a `src/bin/*.rs`, an `examples/*.rs`.
-///
-/// Cargo builds each as its own independent crate, so a name defined in one is
-/// invisible to the other and the two can never actually collide: there is no
-/// ambiguity for a rename to resolve, and W002's "rename one" fix_hint would
-/// be advice to make the code worse. Structural, not name-based — `main` is
-/// merely the most common shared name between two binary targets, not the only
-/// one, and a real copy-paste between them is still caught by W006's body
-/// tiers, which do not care which crate a body lives in.
-///
-/// A pair involving an ordinary library file is NOT exempt: `src/lib.rs` and
-/// everything below it compile into one unit, where a duplicate name is
-/// exactly the ambiguity W002 exists to report.
-pub fn distinct_compilation_units(a: &str, b: &str) -> bool {
-    a != b && is_cargo_binary_root(a) && is_cargo_binary_root(b)
-}
-
-/// Check if a file path is a test file by language convention.
-/// Patterns: *_test.go, test_*.py, *_test.py, conftest.py, *.test.ts,
-/// *.spec.ts, *.test.js, *.spec.js, *_test.rs, *_tests.rs, *_tests_*.rs,
-/// tests.rs
-///
-/// Now that `Definition::in_test_context` marks test symbols precisely per
-/// grammar (Rust/Python/TS in the tree-sitter walk, Go in its Tier-2 pass),
-/// this coarse filename check has narrowed to a fallback: it is the only signal
-/// for a file whose `parse_file` returned no definitions, and a belt-and-braces
-/// safety net for test-support code the per-definition AST rules do not
-/// individually catch — pytest `@pytest.fixture` helpers with non-`test_`
-/// names, Go/TS test utilities. It is also still the sole test signal for the
-/// store-based cross-file filters, whose `GraphNode`s carry no context flags.
-pub fn is_test_file(path: &str) -> bool {
-    let normalized = path.replace('\\', "/");
-
-    // Directory-based: files inside tests/, __tests__/, test/ directories
-    // Check both mid-path ("/tests/") and top-level ("tests/") variants
-    if normalized.contains("/tests/")
-        || normalized.contains("/__tests__/")
-        || normalized.contains("/test/")
-        || normalized.starts_with("tests/")
-        || normalized.starts_with("__tests__/")
-        || normalized.starts_with("test/")
-    {
-        return true;
-    }
-
-    let basename = normalized.rsplit('/').next().unwrap_or(&normalized);
-
-    // Go: *_test.go
-    if basename.ends_with("_test.go") {
-        return true;
-    }
-    // Python: test_*.py, *_test.py, or the pytest fixture-root conftest.py —
-    // fixtures there are consumed via their parameter NAME by tests in other
-    // files, so the graph shows zero callers even though pytest invokes them.
-    if basename == "conftest.py"
-        || (basename.ends_with(".py")
-            && (basename.starts_with("test_") || basename.ends_with("_test.py")))
-    {
-        return true;
-    }
-    // TypeScript/JavaScript: *.test.ts, *.spec.ts, *.test.js, *.spec.js, *.test.tsx, *.spec.tsx
-    if basename.contains(".test.") || basename.contains(".spec.") {
-        return true;
-    }
-    // Rust: *_test.rs, *_tests.rs (plural — this repo's own `#[path =
-    // "..._tests.rs"] mod tests;` convention), or exactly tests.rs.
-    // Suffix match, not substring: "utils_tests.rs" matches, "contests.rs"
-    // does not (no underscore before "tests.rs").
-    // Also `*_tests_*.rs` — the `engine_tests_*.rs` split-file convention
-    // (`engine_tests_e004_misc.rs`, `engine_tests_economy.rs`). `_tests_`
-    // (underscore on both sides) stays precise: "contests_utils.rs" has
-    // "ntests_", no leading underscore, so it does not match.
-    if basename.ends_with("_test.rs")
-        || basename.ends_with("_tests.rs")
-        || basename == "tests.rs"
-        || basename.contains("_tests_")
-    {
-        return true;
-    }
-    false
-}
-
-/// Maximum characters a parenthesized span may cover before it is treated as
-/// unbalanced prose rather than an argument list. Bounds the plan scanner,
-/// which runs over free text where an opening paren may never close.
-pub(crate) const MAX_CLAIM_SPAN: usize = 600;
-
-/// A signature (or a plan's call claim) reduced to the two things keel
-/// compares: how many arguments it takes, and whether it declares a return.
-pub(crate) struct ParsedSig {
-    /// Parameter count with any receiver removed.
-    pub(crate) arity: usize,
-    /// Whether an explicit `-> T` follows the parameter list.
-    pub(crate) has_return: bool,
-    /// Whether a leading `self`/`cls`/`this` receiver was removed from the
-    /// count — the signal that a call site *may* legally spell the receiver as
-    /// its first argument (`Base.__init__(self, x)`, `Rc::clone(&x)`).
-    pub(crate) has_receiver: bool,
-}
-
-/// True for characters that may appear inside an identifier.
-pub(crate) fn is_ident_char(c: char) -> bool {
-    c.is_alphanumeric() || c == '_'
-}
-
-/// True when the `'` at `chars[i]` opens a character literal (`'x'`, `'\n'`)
-/// rather than starting a Rust lifetime (`'a`, `'static`). A literal closes
-/// within two characters (one payload char, or a backslash escape); a lifetime
-/// is `'` + identifier with no closing quote at all. Treating a lifetime as a
-/// string opener swallowed everything up to the *next* tick — in
-/// `(node: Node<'a>, source: &'a [u8])` that includes the top-level comma, so
-/// the arity came out one short and E005 mis-fired on every caller.
-pub(crate) fn tick_opens_literal(chars: &[char], i: usize) -> bool {
-    matches!(chars.get(i + 1), Some('\\')) || matches!(chars.get(i + 2), Some('\''))
-}
-
-/// Index of the `)` closing the `(` at `open`, or `None` when the span is
-/// unbalanced, longer than `MAX_CLAIM_SPAN`, or crosses a blank line.
-pub(crate) fn match_paren(chars: &[char], open: usize) -> Option<usize> {
-    let mut depth = 0i32;
-    let mut in_str: Option<char> = None;
-    let mut prev = '\0';
-    let mut newlines = 0usize;
-    for (offset, &ch) in chars[open..].iter().enumerate() {
-        if offset > MAX_CLAIM_SPAN {
-            return None;
-        }
-        if let Some(q) = in_str {
-            if ch == q && prev != '\\' {
-                in_str = None;
-            }
-            prev = ch;
-            continue;
-        }
-        match ch {
-            '"' | '`' => in_str = Some(ch),
-            '\'' if tick_opens_literal(chars, open + offset) => in_str = Some(ch),
-            '\n' => {
-                newlines += 1;
-                if newlines > 6 {
-                    return None;
-                }
-            }
-            '(' => depth += 1,
-            ')' => {
-                depth -= 1;
-                if depth == 0 {
-                    return Some(open + offset);
-                }
-            }
-            _ => {}
-        }
-        prev = ch;
-    }
-    None
-}
-
-/// Split an argument list on top-level commas, tracking `()`, `[]`, `{}`,
-/// generic `<>` (only when the `<` follows an identifier, so `a < b` is not a
-/// bracket) and string literals. A `'` opens a literal only when it closes as
-/// one ([`tick_opens_literal`]) — a Rust lifetime tick must not swallow the
-/// rest of the list.
-pub(crate) fn split_top_level(args: &str) -> Vec<String> {
-    let chars: Vec<char> = args.chars().collect();
-    let mut parts = Vec::new();
-    let (mut round, mut square, mut curly, mut angle) = (0i32, 0i32, 0i32, 0i32);
-    let mut in_str: Option<char> = None;
-    let mut cur = String::new();
-    let mut prev = '\0';
-    for (i, &ch) in chars.iter().enumerate() {
-        if let Some(q) = in_str {
-            cur.push(ch);
-            if ch == q && prev != '\\' {
-                in_str = None;
-            }
-            prev = ch;
-            continue;
-        }
-        match ch {
-            '"' | '`' => {
-                in_str = Some(ch);
-                cur.push(ch);
-            }
-            '\'' if tick_opens_literal(&chars, i) => {
-                in_str = Some(ch);
-                cur.push(ch);
-            }
-            '(' => {
-                round += 1;
-                cur.push(ch);
-            }
-            ')' => {
-                round -= 1;
-                cur.push(ch);
-            }
-            '[' => {
-                square += 1;
-                cur.push(ch);
-            }
-            ']' => {
-                square -= 1;
-                cur.push(ch);
-            }
-            '{' => {
-                curly += 1;
-                cur.push(ch);
-            }
-            '}' => {
-                curly -= 1;
-                cur.push(ch);
-            }
-            '<' if is_ident_char(prev) => {
-                angle += 1;
-                cur.push(ch);
-            }
-            '>' if angle > 0 && prev != '-' && prev != '=' => {
-                angle -= 1;
-                cur.push(ch);
-            }
-            ',' if round == 0 && square == 0 && curly == 0 && angle == 0 => {
-                parts.push(std::mem::take(&mut cur));
-            }
-            _ => cur.push(ch),
-        }
-        prev = ch;
-    }
-    parts.push(cur);
-    parts
-}
-
-/// Drop a leading receiver parameter (`self`, `&self`, `&mut self`, `cls`,
-/// `this`): it is never written at the call site, so counting it would make
-/// every method call look like it is one argument short. Returns whether a
-/// receiver was actually removed.
-pub(crate) fn strip_receiver(parts: &mut Vec<String>) -> bool {
-    let is_receiver = parts.first().is_some_and(|first| {
-        let t = first
-            .trim()
-            .trim_start_matches('&')
-            .trim()
-            .trim_start_matches("mut ")
-            .trim();
-        let head = t
-            .split(|c: char| c == ':' || c.is_whitespace())
-            .next()
-            .unwrap_or("");
-        matches!(head, "self" | "cls" | "this")
-    });
-    if is_receiver {
-        parts.remove(0);
-    }
-    is_receiver
-}
-
-/// Countable argument count, or `None` when the list is variadic (`*args`,
-/// `...rest`), defaulted (`=`), optional (`?`) or elided (`...`) — all cases
-/// where "the plan says N" and "the code takes N" are not comparable.
-pub(crate) fn countable_arity(parts: &[String]) -> Option<usize> {
-    let trimmed: Vec<&str> = parts.iter().map(|p| p.trim()).collect();
-    if trimmed.len() == 1 && trimmed[0].is_empty() {
-        return Some(0);
-    }
-    for p in &trimmed {
-        if p.is_empty()
-            || p.starts_with('*')
-            || p.starts_with("...")
-            || p.starts_with('…')
-            || p.contains('=')
-            || p.contains('?')
-        {
-            return None;
-        }
-    }
-    Some(trimmed.len())
-}
-
-/// Reduce a signature (`name(params) -> ret`) to arity and return presence.
-/// `None` when there is no parameter list or the parameters are not countable.
-pub(crate) fn parse_signature(sig: &str) -> Option<ParsedSig> {
-    let chars: Vec<char> = sig.chars().collect();
-    let open = chars.iter().position(|&c| c == '(')?;
-    let close = match_paren(&chars, open)?;
-    let args: String = chars[open + 1..close].iter().collect();
-    let mut parts = split_top_level(&args);
-    let has_receiver = strip_receiver(&mut parts);
-    let arity = countable_arity(&parts)?;
-    let tail: String = chars[close + 1..].iter().collect();
-    Some(ParsedSig {
-        arity,
-        has_return: tail.contains("->"),
-        has_receiver,
-    })
-}
-
-/// True when a qualified call name's receiver segment names a *type* rather
-/// than a value — `Base.__init__`, `Rc::clone`, `Self::new` — the one call
-/// shape where the first written argument may legally be the receiver itself.
-///
-/// The receiver is everything before the last `.`/`::`; only its final path
-/// segment is judged (so `mod::Type::method` looks at `Type`). Type-like means
-/// an uppercase initial (the universal type convention in all four languages)
-/// or the literal `Self`. A bare unqualified name has no receiver and is never
-/// type-like.
-pub(crate) fn receiver_is_type_like(call_name: &str) -> bool {
-    let Some(cut) = call_name.rfind("::").max(call_name.rfind('.')) else {
-        return false;
-    };
-    let segment = call_name[..cut].rsplit(['.', ':']).next().unwrap_or("");
-    segment == "Self" || segment.starts_with(|c: char| c.is_uppercase())
-}
-
-/// Strip all whitespace so signatures can be compared ignoring pure
-/// reformatting. A simple "collapse runs of whitespace to one space" would
-/// still miscompare the most common rustfmt/prettier wrap style — one
-/// parameter per line, with the `(`/`)` glued to a newline that has no
-/// corresponding space in the single-line form:
-/// ```text
-/// fn foo(x: i32, y: i32) -> bool
-/// fn foo(
-///     x: i32,
-///     y: i32,
-/// ) -> bool
-/// ```
-/// Collapsing would leave `foo( x: i32...` vs `foo(x: i32...`, still
-/// unequal. Removing whitespace entirely instead of collapsing it handles
-/// this correctly — the token stream is what actually defines the
-/// signature, and whitespace never distinguishes two different signatures.
-pub fn normalize_signature(sig: &str) -> String {
-    sig.chars().filter(|c| !c.is_whitespace()).collect()
-}
-
-/// Split `name` into (start, end) byte ranges for each "word" segment, using
-/// snake_case underscores if present, else camelCase upper-case transitions.
-fn segment_ranges(name: &str) -> Vec<(usize, usize)> {
-    if name.contains('_') {
-        let mut ranges = Vec::new();
-        let mut start = 0;
-        for (i, c) in name.char_indices() {
-            if c == '_' {
-                if i > start {
-                    ranges.push((start, i));
-                }
-                start = i + c.len_utf8();
-            }
-        }
-        if start < name.len() {
-            ranges.push((start, name.len()));
-        }
-        return ranges;
-    }
-
-    // camelCase, with acronym runs treated as one segment: a new segment
-    // starts at an uppercase char only when the previous char is lowercase
-    // (a plain lower->upper transition, e.g. "parse"|"Header"), or when the
-    // previous char is ALSO uppercase but the next char is lowercase (this
-    // is the last capital of a run, which kicks off the next word, e.g. the
-    // second "H" in "HTTPHeader" starts "Header" while "HTTP" stays whole).
-    // Standard camel tokenization: "parseHTTPHeader" -> parse/HTTP/Header,
-    // "toJSONString" -> to/JSON/String, "readIOBuffer" -> read/IO/Buffer.
-    // Without this, consecutive capitals would shatter into one-char
-    // segments ("parseHTTPHeader" -> p-a-r-s-e-H-T-T-P... over-matching
-    // even worse than a single leading segment).
-    let mut ranges = Vec::new();
-    let mut start = 0;
-    let indices: Vec<(usize, char)> = name.char_indices().collect();
-    for i in 1..indices.len() {
-        let (idx, ch) = indices[i];
-        if !ch.is_uppercase() {
-            continue;
-        }
-        let prev_is_upper = indices[i - 1].1.is_uppercase();
-        let next_is_lower = indices.get(i + 1).is_some_and(|&(_, c)| c.is_lowercase());
-        if !prev_is_upper || next_is_lower {
-            ranges.push((start, idx));
-            start = idx;
-        }
-    }
-    if !indices.is_empty() {
-        ranges.push((start, name.len()));
-    }
-    ranges
-}
-
-/// Extract a multi-segment name prefix (e.g. "get_user" from "get_user_name",
-/// "getUser" from "getUserName") used to suggest a placement module.
-///
-/// Only a prefix spanning **at least two** segments is meaningful enough to
-/// suggest a placement — a single leading segment (e.g. "make" from
-/// "make_relative") over-matches wildly, since unrelated functions across
-/// the whole codebase commonly share a first word. Names with fewer than
-/// three total segments can't produce a two-segment prefix while leaving a
-/// remainder (a single word, or a two-word name where the "prefix" would be
-/// the entire name), so they return an empty prefix and never fire W001.
-pub fn extract_prefix(name: &str) -> String {
-    let ranges = segment_ranges(name);
-    if ranges.len() < 3 {
-        return String::new();
-    }
-    name[ranges[0].0..ranges[1].1].to_string()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_fixtures::{definition, file_index, function_node, node_for_definition};
-
-    #[test]
-    fn bind_to_node_prefers_hash_evidence_over_the_name() {
-        // Two stored `run`s; only one carries this def's body. The name alone
-        // would be a coin flip, the hash is decisive.
-        let def = definition("run", "src/a.rs", false);
-        let file = file_index("src/a.rs", vec![def.clone()]);
-        let nodes = vec![
-            function_node(1, "otherhash11", "run", "src/a.rs"),
-            node_for_definition(2, &def),
-        ];
-        assert_eq!(bind_to_node(&def, &file, &nodes).map(|n| n.id), Some(2));
-    }
-
-    #[test]
-    fn bind_to_node_falls_back_only_when_the_name_is_unique_on_both_sides() {
-        // No hash evidence (the stored signature/body moved on), one node and
-        // one definition with the name: the pairing is still forced.
-        let def = definition("run", "src/a.rs", false);
-        let file = file_index("src/a.rs", vec![def.clone()]);
-        let nodes = vec![function_node(1, "otherhash11", "run", "src/a.rs")];
-        assert_eq!(bind_to_node(&def, &file, &nodes).map(|n| n.id), Some(1));
-    }
-
-    #[test]
-    fn bind_to_node_refuses_repeats_on_either_side() {
-        let def = definition("run", "src/a.rs", false);
-        // Repeated in the parse (a free fn beside a method of the same name).
-        let two_defs = file_index("src/a.rs", vec![def.clone(), def.clone()]);
-        let one_node = vec![function_node(1, "otherhash11", "run", "src/a.rs")];
-        assert!(bind_to_node(&def, &two_defs, &one_node).is_none());
-
-        // Repeated in the store (one of the pair was just deleted).
-        let one_def = file_index("src/a.rs", vec![def.clone()]);
-        let two_nodes = vec![
-            function_node(1, "otherhash11", "run", "src/a.rs"),
-            function_node(2, "otherhash22", "run", "src/a.rs"),
-        ];
-        assert!(bind_to_node(&def, &one_def, &two_nodes).is_none());
-    }
-
-    #[test]
-    fn test_is_cargo_binary_root_matches_recognized_roots() {
-        for path in [
-            "build.rs",
-            "crates/x/build.rs",
-            "src/main.rs",
-            "crates/x/src/main.rs",
-            "src/bin/foo.rs",
-            "crates/x/src/bin/foo.rs",
-            "examples/demo.rs",
-            "crates/x/examples/demo.rs",
-        ] {
-            assert!(
-                is_cargo_binary_root(path),
-                "{path} should be a Cargo binary root"
-            );
-        }
-    }
-
-    #[test]
-    fn test_is_cargo_binary_root_rejects_ordinary_source() {
-        for path in [
-            "src/lib.rs",
-            "src/build_rs_helper.rs",
-            "src/mainframe.rs",
-            "crates/x/src/lib.rs",
-        ] {
-            assert!(
-                !is_cargo_binary_root(path),
-                "{path} must not match on substring"
-            );
-        }
-    }
-
-    #[test]
-    fn test_distinct_compilation_units_needs_two_binary_roots() {
-        // Two independent targets: invisible to each other, whatever the name.
-        assert!(distinct_compilation_units("src/bin/a.rs", "src/bin/b.rs"));
-        assert!(distinct_compilation_units("build.rs", "src/main.rs"));
-        assert!(distinct_compilation_units(
-            "examples/one.rs",
-            "crates/x/src/bin/two.rs"
-        ));
-        // A library file on either side shares a namespace — not exempt.
-        assert!(!distinct_compilation_units("src/bin/a.rs", "src/lib.rs"));
-        assert!(!distinct_compilation_units("src/lib.rs", "src/util.rs"));
-        // The same file is one unit, not two.
-        assert!(!distinct_compilation_units("src/bin/a.rs", "src/bin/a.rs"));
-    }
-
-    #[test]
-    fn test_cargo_binary_root_is_rust_only() {
-        // The layout conventions are Cargo's; same-named dirs in other
-        // languages hold ordinary modules that CAN import each other.
-        for path in ["examples/a.ts", "src/bin/cli.ts", "examples/demo.py"] {
-            assert!(!is_cargo_binary_root(path), "{path} is not a Cargo root");
-        }
-        assert!(!distinct_compilation_units(
-            "examples/a.ts",
-            "examples/b.ts"
-        ));
-    }
-
-    #[test]
-    fn test_cargo_binary_root_multi_file_targets_have_one_root() {
-        // dir/<name>/main.rs is the root; its siblings are modules of it.
-        assert!(is_cargo_binary_root("examples/multi/main.rs"));
-        assert!(is_cargo_binary_root("src/bin/tool/main.rs"));
-        for path in [
-            "examples/multi/util.rs",
-            "src/bin/tool/args.rs",
-            "examples/multi/nested/main.rs",
-        ] {
-            assert!(!is_cargo_binary_root(path), "{path} is not a target root");
-        }
-        assert!(!distinct_compilation_units(
-            "examples/multi/main.rs",
-            "examples/multi/util.rs"
-        ));
-    }
-
-    #[test]
-    fn test_src_build_rs_is_a_module_not_a_build_script() {
-        assert!(!is_cargo_binary_root("src/build.rs"));
-        assert!(!is_cargo_binary_root("crates/x/src/build.rs"));
-        assert!(!distinct_compilation_units("src/build.rs", "src/main.rs"));
-    }
-
-    #[test]
-    fn test_parse_signature_counts_params() {
-        // Zero, one, many; nested generic/fn-pointer commas are one param.
-        assert_eq!(parse_signature("fn foo()").unwrap().arity, 0);
-        assert_eq!(parse_signature("fn foo(a: i32, b: str)").unwrap().arity, 2);
-        assert_eq!(parse_signature("def bar(x, y, z)").unwrap().arity, 3);
-        assert_eq!(
-            parse_signature("fn f(m: HashMap<String, i32>, n: u8)")
-                .unwrap()
-                .arity,
-            2
-        );
-        assert_eq!(
-            parse_signature("fn f(cb: fn(a: i32, b: i32) -> i32)")
-                .unwrap()
-                .arity,
-            1
-        );
-        assert!(parse_signature("fn foo").is_none());
-    }
-
-    #[test]
-    fn test_parse_signature_receiver_stripped_and_reported() {
-        // The receiver is never written at an ordinary call site, so it is not
-        // counted — but its presence is reported so E005 can tolerate the
-        // explicit-receiver call shape (`Base.__init__(self, x)`).
-        let sig = parse_signature("fn method(&self, x: i32)").unwrap();
-        assert_eq!(sig.arity, 1);
-        assert!(sig.has_receiver);
-        let sig = parse_signature("def method(self, x)").unwrap();
-        assert_eq!(sig.arity, 1);
-        assert!(sig.has_receiver);
-        let free = parse_signature("fn free(x: i32)").unwrap();
-        assert!(!free.has_receiver);
-    }
-
-    #[test]
-    fn test_parse_signature_rust_lifetimes_do_not_swallow_commas() {
-        // A lifetime tick used to open the "string literal" state, swallowing
-        // everything up to the next tick — including top-level commas. The
-        // repo's own `node_text(node: Node<'a>, source: &'a [u8])` counted as
-        // ONE parameter, so E005 flagged all 19 correct two-argument callers.
-        let sig =
-            parse_signature("node_text(node: tree_sitter::Node<'a>, source: &'a [u8]) -> &'a str")
-                .unwrap();
-        assert_eq!(sig.arity, 2);
-        assert!(sig.has_return);
-        // A lone lifetime (no second tick anywhere) must not unbalance the
-        // paren scan either.
-        let sig = parse_signature("first(x: &'a str)").unwrap();
-        assert_eq!(sig.arity, 1);
-        // Real character literals still open (and close) as literals.
-        let parts = split_top_level("'a', 'b'");
-        assert_eq!(parts.len(), 2);
-        // An escaped-quote char literal keeps its interior comma protected.
-        let parts = split_top_level("f(',', x), y");
-        assert_eq!(parts.len(), 2);
-    }
-
-    #[test]
-    fn test_parse_signature_uncountable_lists_are_none() {
-        // Variadic/defaulted/optional lists are not comparable to a call
-        // site's argument count — E005 must skip them, not guess.
-        assert!(parse_signature("def f(a, b=1)").is_none());
-        assert!(parse_signature("def f(*args)").is_none());
-        assert!(parse_signature("function f(a: number, b?: number)").is_none());
-    }
-
-    #[test]
-    fn test_receiver_is_type_like() {
-        assert!(receiver_is_type_like("Base.__init__"));
-        assert!(receiver_is_type_like("Rc::clone"));
-        assert!(receiver_is_type_like("Self::new"));
-        assert!(receiver_is_type_like("mod::Type::method"));
-        // Value receivers, packages, bare names: not type-like.
-        assert!(!receiver_is_type_like("registry.register"));
-        assert!(!receiver_is_type_like("self.render"));
-        assert!(!receiver_is_type_like("fmt.Println"));
-        assert!(!receiver_is_type_like("plain_call"));
-    }
-
-    #[test]
-    fn test_extract_prefix() {
-        assert_eq!(extract_prefix("x"), "");
-    }
-
-    #[test]
-    fn test_extract_prefix_all_lowercase() {
-        assert_eq!(extract_prefix("process"), "");
-    }
-
-    #[test]
-    fn test_extract_prefix_single_segment_names_never_match() {
-        // Common short verbs/names — never enough segments to produce a prefix.
-        assert_eq!(extract_prefix("run"), "");
-        assert_eq!(extract_prefix("main"), "");
-        assert_eq!(extract_prefix("default"), "");
-    }
-
-    #[test]
-    fn test_extract_prefix_snake_case_multi() {
-        // 3+ segments: prefix is the first TWO segments, not just the first.
-        assert_eq!(extract_prefix("get_user_name"), "get_user");
-        assert_eq!(extract_prefix("get_user_profile_data"), "get_user");
-    }
-
-    #[test]
-    fn test_extract_prefix_camel_case_multi() {
-        assert_eq!(extract_prefix("getUserName"), "getUser");
-        // Only two segments — no match, same rule as snake_case.
-        assert_eq!(extract_prefix("handleRequest"), "");
-    }
-
-    #[test]
-    fn test_extract_prefix_acronym_runs_stay_one_segment() {
-        // Consecutive capitals are one segment (the acronym), not one
-        // segment per letter — otherwise "parseHTTPHeader" would shatter
-        // into p/a/r/s/e/H/T/T/P/... and over-match worse than the old
-        // single-segment behavior. Standard camel tokenization: the run
-        // splits right before its last capital when that capital is
-        // followed by a lowercase letter (it belongs to the next word).
-        assert_eq!(extract_prefix("parseHTTPHeader"), "parseHTTP");
-        assert_eq!(extract_prefix("toJSONString"), "toJSON");
-        assert_eq!(extract_prefix("readIOBuffer"), "readIO");
-
-        // All-caps name: one giant acronym run = a single segment = no
-        // multi-segment prefix possible.
-        assert_eq!(extract_prefix("HTTP"), "");
-    }
-
-    #[test]
-    fn test_extract_prefix_two_segment_names_never_match() {
-        // Regression guard: these used to extract a single-segment prefix
-        // ("make", "process") that over-matched wildly. Now they need a
-        // third segment to leave room for a genuine 2-segment prefix.
-        assert_eq!(extract_prefix("make_relative"), "");
-        assert_eq!(extract_prefix("process_order"), "");
-    }
-
-    #[test]
-    fn test_normalize_signature_whitespace_only_reformat_matches() {
-        // rustfmt's common one-param-per-line wrap style — the `(`/`)`
-        // glue directly to a newline that has no corresponding space in
-        // the single-line form. (A trailing comma, if rustfmt adds one,
-        // is a real content difference, not whitespace — out of scope
-        // for whitespace normalization, and correctly still flagged.)
-        let a = "fn foo(x: i32, y: i32) -> bool";
-        let b = "fn foo(\n    x: i32,\n    y: i32\n) -> bool";
-        assert_eq!(normalize_signature(a), normalize_signature(b));
-
-        // prettier-style wrap without a trailing comma
-        let c = "function foo(\n  x: number,\n  y: number\n): boolean";
-        let d = "function foo(x: number, y: number): boolean";
-        assert_eq!(normalize_signature(c), normalize_signature(d));
-
-        // Extra/irregular spacing
-        assert_eq!(
-            normalize_signature("fn  foo( x:i32 )"),
-            normalize_signature("fn foo(x:i32)")
-        );
-
-        // Leading/trailing whitespace
-        assert_eq!(
-            normalize_signature("  fn foo()  "),
-            normalize_signature("fn foo()")
-        );
-    }
-
-    #[test]
-    fn test_normalize_signature_real_change_differs() {
-        // Added parameter is a real signature change, not just reformatting
-        assert_ne!(
-            normalize_signature("fn foo(x: i32)"),
-            normalize_signature("fn foo(x: i32, y: i32)")
-        );
-        // Type change is a real signature change
-        assert_ne!(
-            normalize_signature("fn foo(x: i32)"),
-            normalize_signature("fn foo(x: i64)")
-        );
-        // Renamed function is a real signature change
-        assert_ne!(
-            normalize_signature("fn foo()"),
-            normalize_signature("fn bar()")
-        );
-    }
-
-    #[test]
-    fn test_is_stub_file() {
-        assert!(is_stub_file("src/models.pyi"));
-        assert!(is_stub_file("types/global.d.ts"));
-        assert!(!is_stub_file("src/models.py"));
-        assert!(!is_stub_file("src/app.ts"));
-    }
-
-    #[test]
-    fn test_is_test_file() {
-        // Go
-        assert!(is_test_file("pkg/handler_test.go"));
-        assert!(!is_test_file("pkg/handler.go"));
-
-        // Python
-        assert!(is_test_file("tests/test_handler.py"));
-        assert!(is_test_file("src/handler_test.py"));
-        assert!(!is_test_file("src/handler.py"));
-        assert!(!is_test_file("src/testing_utils.py")); // not a test file
-
-        // TypeScript/JavaScript
-        assert!(is_test_file("src/handler.test.ts"));
-        assert!(is_test_file("src/handler.spec.ts"));
-        assert!(is_test_file("src/handler.test.js"));
-        assert!(is_test_file("src/handler.spec.tsx"));
-        assert!(!is_test_file("src/handler.ts"));
-
-        // Rust
-        assert!(is_test_file("src/handler_test.rs"));
-        assert!(is_test_file("src/tests.rs"));
-        assert!(!is_test_file("src/handler.rs"));
-
-        // Rust: plural `*_tests.rs` — this repo's own
-        // `#[path = "..._tests.rs"] mod tests;` convention.
-        assert!(is_test_file("crates/keel-core/src/sqlite_tests.rs"));
-        assert!(is_test_file("engine_tests.rs"));
-        assert!(is_test_file("utils_tests.rs"));
-        // Suffix match, not substring: no underscore before "tests.rs".
-        assert!(!is_test_file("contests.rs"));
-        assert!(!is_test_file("src/testing_utils.rs")); // not a test file
-
-        // Rust: `*_tests_*.rs` split-file convention (engine_tests_*.rs) —
-        // full paths and bare basenames.
-        assert!(is_test_file(
-            "crates/keel-enforce/src/engine_tests_e004_misc.rs"
-        ));
-        assert!(is_test_file("engine_tests_economy.rs"));
-        assert!(is_test_file("engine_tests_batch_suppress.rs"));
-        // `_tests_` needs an underscore on BOTH sides: "contests_utils.rs"
-        // has "ntests_" (no leading underscore), so it does not match.
-        assert!(!is_test_file("contests_utils.rs"));
-
-        // Directory-based detection
-        assert!(is_test_file("src/tests/helpers.py"));
-        assert!(is_test_file("project/test/utils.go"));
-        assert!(is_test_file("src/__tests__/handler.ts"));
-        assert!(!is_test_file("src/contest/handler.py")); // "contest" != "test"
-
-        // Top-level test directories (no parent path prefix)
-        assert!(is_test_file("tests/helpers.py"));
-        assert!(is_test_file("test/utils.go"));
-        assert!(is_test_file("__tests__/handler.ts"));
-    }
-
-    #[test]
-    fn test_is_test_file_conftest() {
-        // pytest's fixture-root file: fixtures are consumed by parameter name
-        // in other test files, so it must be recognized as a test file even
-        // though its basename matches none of the test_*/{_test} patterns.
-        assert!(is_test_file("conftest.py"));
-        assert!(is_test_file("app/tests/conftest.py"));
-        // Exact basename match only — a name that merely contains "conftest"
-        // is ordinary production code.
-        assert!(!is_test_file("myconftest.py"));
-    }
-}
